+++ frontend/src/screens/SyllabusView/Overview/Overview.tsx
@@ -9,7 +9,7 @@ import {useNavigate, useLocation} from "react-router-dom";
 import OverviewCard from "./OverviewCard";
 import {loadSyllabusSections, SectionData} from "../../../utils/loadSyllabusSections";
 import AppLayout from "../../../SyllabusLayout/SyllabusPageHeader"
-import RedirectingModal from "../../../components/RedirectingModal/RedirectingModal";
+import RedirectingModal from "../../../components/Modals/RedirectingModal/RedirectingModal";
 import './Overview.css'
 import {handleBack, handleNext} from "../../../components/Button/ButtonLogic";
 import {
